fix : update nft id; replace uint -> uint64
