+++ nft/nft_encode.go
@@ -1,22 +1,10 @@
 package nft
 
 import (
-	extensioncurrency "github.com/ProtoconNet/mitum-currency-extension/currency"
 	"github.com/spikeekips/mitum/base"
 	"github.com/spikeekips/mitum/util"
 	"github.com/spikeekips/mitum/util/encoder"
 )
-
-func (nid *NFTID) unpack(
-	enc encoder.Encoder,
-	collection string,
-	idx uint,
-) error {
-	nid.collection = extensioncurrency.ContractID(collection)
-	nid.idx = idx
-
-	return nil
-}
 
 func (nft *NFT) unpack(
 	enc encoder.Encoder,
